+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 # Changelog
 
 
+## v1.7.1 (2024-06-30)
+
+* **Bug fixes:**
+  * Fixed a bug on `ElasticsearchEx.Sharder` where 0 was excluded from valid number
+
+## v1.7.0 (2024-06-13)
+
+* **New features:**
+  * Added support for Elixir 1.17.0 and Erlang 27.0
+  * Added a module `ElasticsearchEx.Document` to host a document from Elasticsearch
+
+* **Changes:**
+  * Updated dependencies
+  * Fixed `Req` deprecations
+
+## v1.6.0 (2024-05-15)
+
+* **New features:**
+  * Added a module `ElasticsearchEx.Serializer` to serialize document sources
+  * Added a module `ElasticsearchEx.Deserializer` to deserialize document sources
+
 ## v1.5.0 (2024-05-09)
 
 * **Changes:**